Fixed shading (per vertex poly colours) on step 1.5 hardware have the viewport ambient value added to them. This fixes various shading on scud. To do this had to switch the maths to the vertex shader.

+++ Src/Graphics/New3D/PolyHeader.cpp
@@ -344,7 +344,7 @@ UINT8 PolyHeader::Transparency()
 		return 255;					// without this check we get overflow. In the SDK, values are explicitly clamped to 0-32.
 	}
 
-	return (UINT8)((((header[6] >> 18) & 0x3F) * 255) / 32.f);	
+	return (((header[6] >> 18) & 0x3F) * 255) / 32;	
 }
 
 bool PolyHeader::PolyAlpha()
